pyproject.toml
[tool.poetry]
name = "teachbase"
version = "0.1.0"
description = ""
authors = ["Andrew Klyuchnikov <off-rey@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
drf-yasg = "^1.21.3"
psycopg2 = "^2.9.3"
teachbase-api = {git = "https://github.com/AndreyKlychnikov/teachbase_api.git", rev = "main"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
